Moved MotorConroller.jave to comment/hardware
