+++ src/main/java/frc/robot/common/hardware/MotorController.java
@@ -5,12 +5,11 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot;
+package frc.robot.common.hardware;
 
 import com.revrobotics.CANEncoder;
 import com.revrobotics.CANPIDController;
 import com.revrobotics.CANSparkMax;
-import com.revrobotics.SparkMax;
 import com.revrobotics.CANSparkMaxLowLevel.*;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
@@ -26,7 +25,12 @@ public class MotorController {
     private double mP;
     private double mI;
     private double mD;
-
+    
+    /**
+    * Creates a SparkMax and resets it to default.
+    * @param name the prefix for all the variables in SmartDashboard 
+    * @param deviceID the Device ID for the new SparkMax
+    */
     public MotorController(String name, int deviceID) {
         mName = name;
         mSparkMax = new CANSparkMax(deviceID, MotorType.kBrushless);
@@ -37,6 +41,14 @@ public class MotorController {
         mSparkMax.restoreFactoryDefaults();
     }
 
+    /**
+     * Creates the PID for the controller and calls motor controller
+     * @see MotorController(String name, int deviceID)
+     * @param name the prefix for all the variables in SmartDashboard 
+     * @param deviceID the Device ID for the new SparkMax
+     * @param smartCurrentLimit the current limit for prevention of brown outs
+     * @param enablePid finds whether we are enabling pid
+     */
     //Todo: we must add constants for PID values once we have found the right values
     public MotorController(String name, int deviceID, int smartCurrentLimit, boolean... enablePid) {
         this(name, deviceID); //calls the constructor for correct arguements
@@ -54,18 +66,33 @@ public class MotorController {
         mSparkMax.setOpenLoopRampRate(.1);
     }
 
+    /**
+     * Gets the Spark Max
+     * @return returns Spark Max
+     */
     public CANSparkMax getSparkMax() {
         return mSparkMax;
     }
 
+    /**
+     * Gets the Encoder
+     * @return returns Encoder
+     */
     public CANEncoder getEncoder() {
         return mEncoder;
     }
 
+    /**
+     * Gets the PID
+     * @return returns PID controller
+     */
     public CANPIDController getPID() {
         return mPIDController;
     }
 
+    /**
+     * Sets the PID for the controller and puts the PID values in Smart Dashboard
+     */
     public void setPID() {
         mPIDController.setP(mP);
         mPIDController.setI(mI);
@@ -75,12 +102,11 @@ public class MotorController {
         SmartDashboard.putNumber(mName+" D Value", mD);
     }
 
+    /**
+     * Updates the Smart Dashboard and checks the PID values to determine if update is needed
+     */
     public void updateSmartDashboard() {
         //The simulation crashes whenever .getEncoder() is called
-        if(Robot.isReal()) {
-            SmartDashboard.putNumber(mName + " Encoder Position", mEncoder.getPosition());
-            SmartDashboard.putNumber(mName + " Encoder Velocity", mEncoder.getVelocity());
-        }
         if(mPIDController != null) {
             if (SmartDashboard.getNumber(mName + " P Value", mP) != mP) {
                 mP = SmartDashboard.getNumber(mName + " P Value", mP);
