+++ src/main/java/frc/robot/subsystems/HopperSubsystem.java
@@ -12,7 +12,7 @@ import com.revrobotics.ControlType;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.MotorController;
+import frc.robot.common.hardware.MotorController;
 
 public class HopperSubsystem extends SubsystemBase {
   /**
