+++ src/main/java/frc/robot/subsystems/IntakeSubsystem.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.DoubleSolenoid.Value;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.MotorController;
+import frc.robot.common.hardware.MotorController;
 
 public class IntakeSubsystem extends SubsystemBase {
   /**
