+++ src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -8,7 +8,6 @@
 package frc.robot.subsystems;
 
 import frc.robot.Constants;
-import frc.robot.MotorController;
 import frc.robot.Game;
 import com.revrobotics.CANPIDController;
 import com.revrobotics.ControlType;
@@ -23,10 +22,9 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Game;
-    
-import frc.robot.MotorController;
 import frc.robot.Robot;
 import frc.robot.common.hardware.Limelight;
+import frc.robot.common.hardware.MotorController;
 
 public class ShooterSubsystem extends SubsystemBase {
   /**
